requirements.txt
absl-py==0.15.0
astunparse==1.6.3
cached-property==1.5.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
flatbuffers==1.12
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.4
importlib-metadata==6.0.0
install==1.3.5
Keras==2.3.0
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.4.1
MarkupSafe==2.1.2
numpy==1.19.5
oauthlib==3.2.2
opt-einsum==3.3.0
Pillow==9.0.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.7.3
six==1.15.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.3
tensorflow-estimator==2.5.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.14
Werkzeug==2.2.2
wrapt==1.12.1
zipp==3.12.0
